+++ node/src/BaseClient.ts
@@ -80,6 +80,7 @@ import {
     createGet,
     createGetBit,
     createGetDel,
+    createGetEx,
     createGetRange,
     createHDel,
     createHExists,
@@ -196,6 +197,7 @@ import {
     createZScore,
     StreamClaimOptions,
     createXClaim,
+    GetExOptions,
 } from "./Commands";
 import {
     ClosingError,
@@ -837,6 +839,25 @@ export class BaseClient {
      */
     public get(key: string): Promise<string | null> {
         return this.createWritePromise(createGet(key));
+    }
+
+    /** Get the value of `key` and optionally set its expiration. `GETEX` is similar to `GET`.
+     * See https://valkey.io/commands/getex for more details.
+     *
+     * @param key - The key to retrieve from the database.
+     * @param options - (Optional) set expiriation to the given key.
+     *                  Equivalent to [`EX` | `PX` | `EXAT` | `PXAT` | `PERSIST`] in the VALKEY API.
+     * @returns If `key` exists, returns the value of `key` as a `string`. Otherwise, return `null`.
+     *
+     * @example
+     * ```typescript
+     * // Example usage of get method to retrieve the value of a key
+     * const result = await client.getex("key", {expiry: { type: "seconds", count: 5 }});
+     * console.log(result); // Output: 'value'
+     * ```
+     */
+    public getex(key: string, options?: GetExOptions): Promise<string | null> {
+        return this.createWritePromise(createGetEx(key, options));
     }
 
     /**